@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:parse_server_sdk_flutter/parse_server_sdk.dart';
 import 'package:precept_back4app_backend/backend/back4app/authenticator/authenticator.dart';
 import 'package:precept_back4app_backend/backend/back4app/dataProvider/pBack4AppDataProvider.dart';
@@ -8,7 +7,7 @@ import 'package:precept_backend/backend/user/authenticator.dart';
 import 'package:precept_script/data/provider/documentId.dart';
 
 class Back4AppDataProvider extends DataProvider<PBack4AppDataProvider> {
-  Back4AppDataProvider({@required PBack4AppDataProvider config}) : super(config: config);
+  Back4AppDataProvider({required PBack4AppDataProvider config}) : super(config: config);
 
   @override
   Authenticator<PBack4AppDataProvider, ParseUser> createAuthenticator(
@@ -27,6 +26,7 @@ class Back4AppDataProvider extends DataProvider<PBack4AppDataProvider> {
 class Back4App {
   static register() {
     dataProviderLibrary.register(
-        config: PBack4AppDataProvider, builder: (config) => Back4AppDataProvider(config: config));
+        config: PBack4AppDataProvider,
+        builder: (config) => Back4AppDataProvider(config: config as PBack4AppDataProvider));
   }
 }
